refactor: the number 2 does not constitute a new word
