+++ source/lib/geom/vec2.ts
